pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "fastapi>=0.116.1",
    "fastapi-mcp>=0.3.7",
    "flagembedding>=1.3.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-elasticsearch==0.3.2",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.6",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.5.4",
    "python-dotenv>=1.1.1",
    "python-json-logger>=3.3.0",
    "uvicorn>=0.35.0",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
